fix: namespace validation
Assisted by: Cursor
