+++ controllers/argocd/util.go
@@ -1218,31 +1218,6 @@ func (r *ReconcileArgoCD) triggerRollout(obj interface{}, key string) error {
 	}
 }
 
-func allowedNamespace(current string, namespaces string) bool {
-
-	clusterConfigNamespaces := splitList(namespaces)
-	if len(clusterConfigNamespaces) > 0 {
-		if clusterConfigNamespaces[0] == "*" {
-			return true
-		}
-
-		for _, n := range clusterConfigNamespaces {
-			if n == current {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-func splitList(s string) []string {
-	elems := strings.Split(s, ",")
-	for i := range elems {
-		elems[i] = strings.TrimSpace(elems[i])
-	}
-	return elems
-}
-
 func containsString(arr []string, s string) bool {
 	for _, val := range arr {
 		if strings.TrimSpace(val) == s {
